requirements.txt
certifi==2019.3.9
chardet==3.0.4
idna==2.8
jsonpickle==1.1
PyNamecheap==0.0.3
python-digitalocean==1.14.0
requests==2.21.0
urllib3==1.24.1
